requirements.txt
discord.py==2.3.2
pandas==1.4.3
python-dotenv==1.0.0
